requirements.txt
prometheus-client>=0.12.0
pyyaml>=6.0
requests
pysolarmanv5
